+++ R/rasterize.r
@@ -6,12 +6,13 @@
 #'
 #' @param y A `GRaster`: The new raster will have the same extent and resolution as this raster.
 #'
-#' @param background Numeric or `NA` (default): Value to put in cells that are not covered by the `GVector`.
+#' @param field Character: Name of a column in the data table of `y` to "burn" into the raster. If not `""` (default), then the output will be a [categorical][tutorial_raster_data_types] raster. If `field` is `""`, then all geometries will be "burned" to the raster and have the same value.
 #'
-#' @param byGeom,collapse Logical:
-#' * If `byGeom` is `TRUE` and `collapse` is `TRUE`, the output will have just one raster layer, but with values that correspond to each geometry (i.e., each geometry will get a different value). If geometries overlap, then category value of the geometry on "top" will be assigned to the relevant cells. 
-#' * If `byGeom` is `TRUE`, and `collapse` is `FALSE`, then create a separate raster layer for each geometry in `x`.
-#' * If `byGeom` is `FALSE`, then create a single raster layer with the same value for all geometries. Argument `collapse` is ignored.
+#' @param background Numeric or `NA` (default): Value to put in cells that are not covered by the `GVector`. Note that if this is not `NA` and not an integer, then the output cannot be a categorical raster (i.e., there will be no "levels" table associated with it).
+#'
+#' @param by Either `NULL` (default) or character: If this is not `NULL`, then the `GVector` will be subset by the values in the field named by `by`. The output will be a multi-layer raster, with one layer per unique value in `by`.
+#'
+#' @param verbose Logical: If `by` is not `NULL`, display progress.
 #'
 #' @returns A `GRaster`.
 #'
@@ -28,96 +29,108 @@ methods::setMethod(
 	function(
 		x,
 		y,
+		field = "",
 		background = NA,
-		byGeom = FALSE,
-		collapse = TRUE
+		by = NULL,
+		verbose = TRUE
 	) {
 	
 	compareGeom(x, y)
 	.locationRestore(x)
 	.region(y)
 
-	gtype <- geomtype(x, grass = TRUE)
-	src <- .rasterize(x = x, y = y, background = background, byGeom = byGeom, collapse = collapse)
-	.makeGRaster(src)
+	out <- .rasterize(x = x, y = y, field = field, background = background, by = by, verbose = verbose)
+	.makeGRaster(out$src, levels = out$levels)
 
 	} # EOF
 )
 
 #' @param x [sources()] name of a `GVector`.
 #' @param y [sources()] name of a `GRaster`
+#' @param field Name of a field in the data table of `x` (can be `""`).
 #' @param background Numeric or `NA`.
-#' @param byGeom Logical.
-#' @param collapse Logical.
+#' @param by `NULL` or name of a field in the data table of `x`.
+#' @param verbose Logical.
+#'
+#' @returns A `list` with the [sources()] name of the output raster, plus a `levels` table (can be `NULL`).
 #'
 #' @noRd
-.rasterize <- function(x, y, background, byGeom, collapse) {
+.rasterize <- function(x, y, field, background, by, verbose) {
 
 	### create different raster layer for each geometry
-	if (byGeom) {
+	if (!is.null(by)) {
+
+		bys <- unique(x@table[[by]])
+		nBys <- length(bys)
+		src <- rep(NA_character_, nBys)
+		levels <- list()
+		if (verbose & nBys > 1L) pb <- utils::txtProgressBar(min = 0, max = nBys, initial = 0, style = 2)
+
+		for (i in seq_len(nBys)) {
+
+			if (verbose) {
+				utils::setTxtProgressBar(pb, i)
+				utils::flush.console()
+			}
+
+			index <- which(x@table[[by]] == bys[i])
+			xx <- x[index]
+
+			thisOut <- .rasterize(xx, y, field = field, background = background, by = NULL, verbose = FALSE)
+			src[i] <- thisOut$src
+			levels[[i]] <- thisOut$levels
+		
+		}
+		if (verbose) close(pb)
+
+	} else {
+
+		gtype <- geomtype(x, grass = TRUE)
+		src <- .makeSourceName("rasterize_v_to_rast", "raster")
 
 		# if by geometry but burned to the same raster
-		if (collapse) {
+		if (field == "") {
 
-			gtype <- geomtype(x, grass = TRUE)
-			src <- .makeSourceName("rasterize_v_to_rast", "raster")
+			levels <- data.table::data.table(NULL)
+
 			args <- list(
 				cmd = "v.to.rast",
 				input = sources(x),
 				output = src,
-				use = "cat",
+				use = "val",
+				value = 1,
+				type = gtype,
+				memory = faster("memory"),
+				flags = c(.quiet(), "overwrite")
+			)
+		
+		} else {
+			
+			# rasterize by field in data table
+			cats <- .vCats(x)
+			uniCats <- sort(unique(cats))
+			fieldVals <- x@table[[field]]
+			renums <- omnibus::renumSeq(fieldVals)
+			levels <- data.table::data.table(value = unique(renums), DUMMYDUMMY_ = unique(fieldVals))
+			names(levels)[2L] <- field
+			levels <- levels[order(value)]
+			db <- data.table::data.table(cat = uniCats, DUMMYDUMMY_ = renums)
+			.vAttachDatabase(x, db)
+
+			args <- list(
+				cmd = "v.to.rast",
+				input = sources(x),
+				output = src,
+				use = "attr",
+				attribute_column = "DUMMYDUMMY_",
 				type = gtype,
 				memory = faster("memory"),
 				flags = c(.quiet(), "overwrite")
 			)
 
-			if (gtype == "line") args$flags <- c(args$flags, "d")
-
-			do.call(rgrass::execGRASS, args = args)
-			out <- .makeGRaster(src, "rasterize")
-
-		} else {
-			
-			ng <- ngeom(x)
-			for (i in seq_len(ng)) {
-			
-				xx <- x[i]
-
-				thisOut <- rasterize(
-					x = xx, y = y,
-					background = background,
-					byGeom = FALSE
-				)
-
-				if (i == 1L) {
-					out <- thisOut
-				} else {
-					out <- c(out, thisOut)
-				}
-
-			} # next geometry
-
-
-		} # if one raster per geometry, not collapsed
-
-	### all geometries at once
-	} else {
-
-		gtype <- geomtype(x, grass = TRUE)
-		src <- .makeSourceName("rasterize_v_to_rast", "raster")
-		args <- list(
-			cmd = "v.to.rast",
-			input = sources(x),
-			output = src,
-			use = "val",
-			value = 1,
-			type = gtype,
-			memory = faster("memory"),
-			flags = c(.quiet(), "overwrite")
-		)
+		}
 
 		if (gtype == "line") args$flags <- c(args$flags, "d")
-
 		do.call(rgrass::execGRASS, args = args)
 
 		if (!is.na(background)) {
@@ -132,9 +145,28 @@ methods::setMethod(
 				flags = c(.quiet(), "overwrite")
 			)
 
+			if (nrow(levels) > 0L) {
+			
+				if (is.integer(background) | omnibus::is.wholeNumber(background)) {
+
+					background <- as.integer(background)
+
+					levAdd <- data.table::data.table(value = background, DUMMYDUMMY_ = "background")
+					names(levAdd) <- names(levels)
+					levels <- rbind(levels, levAdd)
+					levels <- levels[!duplicated(levels)]
+					levels <- levels[order(value)]
+
+				} else {
+					levels <- data.table::data.table(NULL)
+				}
+				
+			}
+
 		}
 
 	}
-	src
+
+	list(src = src, levels = levels)
 
 }